refactor: CLI explore
