+++ packages/daf-cli/src/explore/managed-identities.ts
